+++ liblogging/src/datastore/auth/heartbeat.rs
@@ -9,7 +9,6 @@ pub async fn routine(auth_handler: AuthHandler) {
             .heartbeat(auth_handler.id.clone(), auth_handler.secret.clone())
             .await
         else {
-            log::warn!("Failed to send heartbeat to the server. Retrying in 10 seconds...");
             tokio::time::sleep(Duration::from_secs(10)).await;
             continue;
         };
@@ -19,31 +18,5 @@ pub async fn routine(auth_handler: AuthHandler) {
             *t = heartbeat_response.token();
             drop(t);
         }
-    }
-}
-
-pub(crate) enum GenericHeartbeatResponse {
-    AppGuard(nullnet_libappguard::HeartbeatResponse),
-    WallGuard(nullnet_libwallguard::HeartbeatResponse),
-}
-
-impl GenericHeartbeatResponse {
-    fn token(&self) -> String {
-        match self {
-            GenericHeartbeatResponse::AppGuard(response) => response.token.clone(),
-            GenericHeartbeatResponse::WallGuard(response) => response.token.clone(),
-        }
-    }
-}
-
-impl From<nullnet_libappguard::HeartbeatResponse> for GenericHeartbeatResponse {
-    fn from(val: nullnet_libappguard::HeartbeatResponse) -> Self {
-        GenericHeartbeatResponse::AppGuard(val)
-    }
-}
-
-impl From<nullnet_libwallguard::HeartbeatResponse> for GenericHeartbeatResponse {
-    fn from(val: nullnet_libwallguard::HeartbeatResponse) -> Self {
-        GenericHeartbeatResponse::WallGuard(val)
     }
 }
