+++ liblogging/src/datastore/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod auth;
 pub(crate) mod config;
+pub(crate) mod generic_log;
+pub(crate) mod grpc_interface;
 pub(crate) mod transmitter;
 pub(crate) mod wrapper;
