+++ liblogging/src/datastore/wrapper.rs
@@ -1,5 +1,6 @@
 use crate::datastore::auth::{AuthHandler, GrpcInterface};
 use crate::datastore::config::DatastoreConfig;
+use crate::datastore::generic_log::GenericLog;
 
 pub(crate) struct ServerWrapper {
     inner: GrpcInterface,
@@ -18,32 +19,5 @@ impl ServerWrapper {
         let token = self.auth.get_token().await;
 
         self.inner.handle_logs(token, logs).await
-    }
-}
-
-#[derive(Clone)]
-pub(crate) struct GenericLog {
-    pub(crate) timestamp: String,
-    pub(crate) level: String,
-    pub(crate) message: String,
-}
-
-impl From<GenericLog> for nullnet_libappguard::Log {
-    fn from(val: GenericLog) -> nullnet_libappguard::Log {
-        nullnet_libappguard::Log {
-            timestamp: val.timestamp,
-            level: val.level,
-            message: val.message,
-        }
-    }
-}
-
-impl From<GenericLog> for nullnet_libwallguard::Log {
-    fn from(val: GenericLog) -> nullnet_libwallguard::Log {
-        nullnet_libwallguard::Log {
-            timestamp: val.timestamp,
-            level: val.level,
-            message: val.message,
-        }
     }
 }
